requirements.txt
faiss-cpu
langchain_community
langchain
youtube-transcript-api
requests
python-dotenv
mistralai
sentence-transformers
langchain_huggingface
langchain_core
streamlit
fastapi
uvicorn
fastapi